Add files via upload
The project starts with an exploratory analysis to check the relationship between various parameters such as Age, Alcoholism, Day of the Week, Appointment Data, Gender with Target Variable which is "ShowUp and No Showup". Later I have built two predictive models in  Logistic Regression and Decision Tree model to get the significant variables and target prediction on test data.